# docs/swagger.yaml
basePath: /
definitions:
...
  entity.Comment:
    properties:
      created_at:
        type: string
      id:
        type: integer
      post:
        $ref: '#/definitions/entity.Post'
      post_id:
        type: integer
      problem:
        $ref: '#/definitions/entity.Problem'
      problem_id:
        type: integer
      replies:
        description: Relations
        items:
          $ref: '#/definitions/entity.Comment'
        type: array
      reply:
        $ref: '#/definitions/entity.Comment'
      reply_id:
        type: integer
      submission:
        $ref: '#/definitions/entity.Submission'
      submission_id:
        type: integer
      text:
        type: string
      track:
        $ref: '#/definitions/entity.Track'
      track_id:
        type: integer
      updated_at:
        type: string
      user:
        $ref: '#/definitions/entity.User'
      user_id:
        type: integer
      votes:
        items:
          $ref: '#/definitions/entity.Vote'
        type: array
    type: object
  entity.Contest:
    properties:
      created_at:
        type: string
      id:
        type: integer
      link:
        type: string
      link2:
        type: string
      link3:
        type: string
      name:
        type: string
      problem_count:
        type: integer
      problems:
        description: Relations
        items:
          $ref: '#/definitions/entity.Problem'
        type: array
      super_group:
        $ref: '#/definitions/entity.SuperGroup'
      super_group_id:
        type: integer
      type:
        type: string
      unrated:
        type: boolean
      updated_at:
        type: string
      votes:
        description: |-
          Ratings       []Rating       `json:"ratings,omitempty" gorm:"foreignKey:ContestID"`
          DivisionUsers []DivisionUser `json:"division_users,omitempty" gorm:"foreignKey:ContestID"`
        items:
          $ref: '#/definitions/entity.Vote'
        type: array
    type: object
...
  entity.DailyProblem:
    properties:
      created_at:
        type: string
      for_date:
        type: string
      id:
        type: integer
      problem:
        $ref: '#/definitions/entity.Problem'
      problem_id:
        type: integer
      super_group:
        $ref: '#/definitions/entity.SuperGroup'
      super_group_id:
        type: integer
      updated_at:
        type: string
    type: object
  entity.Exercise:
    properties:
      created_at:
        type: string
      group:
        $ref: '#/definitions/entity.Group'
      group_id:
        type: integer
      id:
        type: integer
      problem:
        $ref: '#/definitions/entity.Problem'
      problem_id:
        type: integer
      track:
        $ref: '#/definitions/entity.Track'
      track_id:
        type: integer
      updated_at:
        type: string
    type: object
...
  entity.Group:
    properties:
      country:
        $ref: '#/definitions/entity.Country'
      country_id:
        type: integer
      created_at:
        type: string
      description:
        type: string
      exercises:
        items:
          $ref: '#/definitions/entity.Exercise'
        type: array
      google_oauths:
        items:
          $ref: '#/definitions/entity.GoogleOAuth'
        type: array
      group_sessions:
        items:
          $ref: '#/definitions/entity.GroupSession'
        type: array
      hoa_id:
        description: Head of Academy ID
        type: integer
      hoas:
        items:
          $ref: '#/definitions/entity.HOA'
        type: array
      id:
        type: integer
      invites:
        items:
          $ref: '#/definitions/entity.Invite'
        type: array
      name:
        type: string
      short_name:
        type: string
      super_to_groups:
        items:
          $ref: '#/definitions/entity.SuperToGroup'
        type: array
      updated_at:
        type: string
      users:
        description: Relations
        items:
          $ref: '#/definitions/entity.User'
        type: array
    type: object
...
  entity.Post:
    properties:
      body:
        type: string
      comments:
        description: Relations
        items:
          $ref: '#/definitions/entity.Comment'
        type: array
      created_at:
        type: string
      id:
        type: integer
      post_to_tags:
        items:
          $ref: '#/definitions/entity.PostToTag'
        type: array
      title:
        type: string
      updated_at:
        type: string
      user:
        $ref: '#/definitions/entity.User'
      user_id:
        type: integer
      votes:
        items:
          $ref: '#/definitions/entity.Vote'
        type: array
    type: object
...
  entity.Problem:
    properties:
      comments:
        items:
          $ref: '#/definitions/entity.Comment'
        type: array
      contest:
        $ref: '#/definitions/entity.Contest'
      contest_id:
        type: integer
      created_at:
        type: string
      daily_problems:
        items:
          $ref: '#/definitions/entity.DailyProblem'
        type: array
      difficulty:
        type: string
      exercises:
        items:
          $ref: '#/definitions/entity.Exercise'
        type: array
      id:
        type: integer
      link:
        type: string
      name:
        type: string
      platform:
        type: string
      problem_tracks:
        items:
          $ref: '#/definitions/entity.ProblemTrack'
        type: array
      submissions:
        description: Relations
        items:
          $ref: '#/definitions/entity.Submission'
        type: array
      tag:
        type: string
      track:
        $ref: '#/definitions/entity.Track'
      track_id:
        type: integer
      updated_at:
        type: string
      votes:
        items:
          $ref: '#/definitions/entity.Vote'
        type: array
    type: object
  entity.ProblemTrack:
    properties:
      created_at:
        type: string
      id:
        type: integer
      problem:
        $ref: '#/definitions/entity.Problem'
      problem_id:
        type: integer
      track:
        $ref: '#/definitions/entity.Track'
      track_id:
        type: integer
      updated_at:
        type: string
    type: object
  entity.RecentAction:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      type:
        type: string
      updated_at:
        type: string
      user:
        $ref: '#/definitions/entity.User'
      user_id:
        type: integer
    type: object
...
  entity.Submission:
    properties:
      code:
        type: string
      comments:
        description: Relations
        items:
          $ref: '#/definitions/entity.Comment'
        type: array
      created_at:
        type: string
      id:
        type: integer
      in_contest:
        description: Whether it was solved in a contest
        type: integer
      language:
        type: string
      problem:
        $ref: '#/definitions/entity.Problem'
      problem_id:
        type: integer
      time_spent:
        description: Time spent in seconds
        type: integer
      tries:
        description: Number of attempts
        type: integer
      updated_at:
        type: string
      user:
        $ref: '#/definitions/entity.User'
      user_id:
        type: integer
      verified:
        type: boolean
      votes:
        items:
          $ref: '#/definitions/entity.Vote'
        type: array
    type: object
  entity.SuperGroup:
    properties:
      contests:
        items:
          $ref: '#/definitions/entity.Contest'
        type: array
      created_at:
        type: string
      daily_problems:
        items:
          $ref: '#/definitions/entity.DailyProblem'
        type: array
      id:
        type: integer
      name:
        type: string
      super_to_groups:
        items:
          $ref: '#/definitions/entity.SuperToGroup'
        type: array
      tracks:
        description: Relations
        items:
          $ref: '#/definitions/entity.Track'
        type: array
      updated_at:
        type: string
    type: object
...
  entity.Track:
    properties:
      active:
        type: boolean
      comments:
        items:
          $ref: '#/definitions/entity.Comment'
        type: array
      created_at:
        type: string
      exercises:
        items:
          $ref: '#/definitions/entity.Exercise'
        type: array
      id:
        type: integer
      name:
        type: string
      problem_tracks:
        items:
          $ref: '#/definitions/entity.ProblemTrack'
        type: array
      problems:
        description: Relations
        items:
          $ref: '#/definitions/entity.Problem'
        type: array
      super_group:
        $ref: '#/definitions/entity.SuperGroup'
      super_group_id:
        type: integer
      updated_at:
        type: string
      votes:
        items:
          $ref: '#/definitions/entity.Vote'
        type: array
    type: object
  entity.User:
    properties:
      api_tokens:
        items:
          $ref: '#/definitions/entity.APIToken'
        type: array
      attendances:
        items:
          $ref: '#/definitions/entity.Attendance'
        type: array
      birthday:
        description: Personal Information
        type: string
      code_of_conduct:
        type: string
      codeforces:
        type: string
      comments:
        items:
          $ref: '#/definitions/entity.Comment'
        type: array
      config:
        type: string
      country:
        $ref: '#/definitions/entity.Country'
      country_id:
        type: integer
      created_at:
        description: Timestamps
        type: string
      cv:
        description: Professional Details
        type: string
      department:
        type: string
      email:
        type: string
      expected_graduation_date:
        type: string
      gender:
        type: string
      github:
        type: string
      group:
        $ref: '#/definitions/entity.Group'
      group_id:
        type: integer
      hackerrank:
        type: string
      head_attendances:
        items:
          $ref: '#/definitions/entity.Attendance'
        type: array
      id:
        description: Core Identity Fields
        type: integer
      inactive:
        type: boolean
      instagram:
        type: string
      joined_date:
        type: string
      leetcode:
        description: |-
          Coding Profiles (optional)
          Use pointer types so that if no value is provided, they remain nil.
        type: string
      linkedin:
        description: Social Media (all optional)
        type: string
      mentor_name:
        type: string
      name:
        type: string
      phone:
        description: Contact Information
        type: string
      photo:
        description: System Fields
        type: string
      posts:
        items:
          $ref: '#/definitions/entity.Post'
        type: array
      preferred_language:
        type: string
      recent_actions:
        items:
          $ref: '#/definitions/entity.RecentAction'
        type: array
      role:
        $ref: '#/definitions/entity.Role'
      role_id:
        description: Role and Relationships
        type: integer
      short_bio:
        type: string
      student_id:
        type: string
      submissions:
        description: Relations (using GORM associations)
        items:
          $ref: '#/definitions/entity.Submission'
        type: array
      telegram_uid:
        type: string
      telegram_username:
        type: string
      tshirt_color:
        description: Physical Attributes
        type: string
      tshirt_size:
        type: string
      university:
        description: Academic Information
        type: string
      updated_at:
        type: string
    type: object
  entity.Vote:
    properties:
      comment:
        $ref: '#/definitions/entity.Comment'
      comment_id:
        type: integer
      contest:
        $ref: '#/definitions/entity.Contest'
      contest_id:
        type: integer
      created_at:
        type: string
      id:
        type: integer
      post:
        $ref: '#/definitions/entity.Post'
      post_id:
        type: integer
      problem:
        $ref: '#/definitions/entity.Problem'
      problem_id:
        type: integer
      submission:
        $ref: '#/definitions/entity.Submission'
      submission_id:
        type: integer
      track:
        $ref: '#/definitions/entity.Track'
      track_id:
        type: integer
      type:
        description: e.g., 1=upvote, -1=downvote
        type: integer
      updated_at:
        type: string
      user:
        $ref: '#/definitions/entity.User'
      user_id:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@hub.a2sv.org
    name: API Support
  description: This is the API documentation for the Hub backend.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Hub API
  version: "1.0"
paths:
...
  /api/problems:
    get:
      description: Get a list of all problems
      produces:
      - application/json
      responses:
        "200":
          description: List of problems
          schema:
            items:
              $ref: '#/definitions/entity.Problem'
            type: array
      summary: List problems
      tags:
      - Problems
    post:
      consumes:
      - application/json
      description: Create a new problem entry
      parameters:
      - description: Problem data
        in: body
        name: problem
        required: true
        schema:
          $ref: '#/definitions/entity.Problem'
      produces:
      - application/json
      responses:
        "201":
          description: Problem created successfully
          schema:
            $ref: '#/definitions/entity.Problem'
        "400":
          description: Invalid request body
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a new problem
      tags:
      - Problems
  /api/problems/{id}:
    delete:
      description: Delete a problem by its ID
      parameters:
      - description: Problem ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Problem deleted successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid problem ID
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Problem not found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete a problem
      tags:
      - Problems
    get:
      description: Get a problem by its ID
      parameters:
      - description: Problem ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Problem details
          schema:
            $ref: '#/definitions/entity.Problem'
        "400":
          description: Invalid problem ID
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Problem not found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get problem by ID
      tags:
      - Problems
    patch:
      consumes:
      - application/json
      description: Update a problem by its ID
      parameters:
      - description: Problem ID
        in: path
        name: id
        required: true
        type: integer
      - description: Problem data
        in: body
        name: problem
        required: true
        schema:
          $ref: '#/definitions/entity.Problem'
      produces:
      - application/json
      responses:
        "200":
          description: Problem updated successfully
          schema:
            $ref: '#/definitions/entity.Problem'
        "400":
          description: Invalid input
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update a problem
      tags:
      - Problems
  /api/problems/name/{name}:
    get:
      description: Get a problem by its name
      parameters:
      - description: Problem name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Problem details
          schema:
            $ref: '#/definitions/entity.Problem'
        "400":
          description: Invalid problem name
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Problem not found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get problem by name
      tags:
      - Problems
  /api/recent-actions:
    get:
      description: Get a list of all recent user actions
      produces:
      - application/json
      responses:
        "200":
          description: List of recent actions
          schema:
            items:
              $ref: '#/definitions/entity.RecentAction'
            type: array
      summary: List recent actions
      tags:
      - RecentActions
    post:
      consumes:
      - application/json
      description: Create a new recent user action
      parameters:
      - description: Recent action details
        in: body
        name: recentAction
        required: true
        schema:
          $ref: '#/definitions/entity.RecentAction'
      produces:
      - application/json
      responses:
        "201":
          description: Recent action created
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid request body
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create recent action
      tags:
      - RecentActions
  /api/recent-actions/{id}:
    delete:
      description: Delete a recent user action
      parameters:
      - description: Recent Action ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Recent action deleted
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid ID
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete recent action
      tags:
      - RecentActions
    get:
      description: Get details of a recent action by its ID
      parameters:
      - description: Recent Action ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Recent action details
          schema:
            $ref: '#/definitions/entity.RecentAction'
        "400":
          description: Invalid ID
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get recent action by ID
      tags:
      - RecentActions
    put:
      consumes:
      - application/json
      description: Update a recent user action
      parameters:
      - description: Recent Action ID
        in: path
        name: id
        required: true
        type: integer
      - description: Recent action details
        in: body
        name: recentAction
        required: true
        schema:
          $ref: '#/definitions/entity.RecentAction'
      produces:
      - application/json
      responses:
        "200":
          description: Recent action updated
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid request body
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update recent action
      tags:
      - RecentActions
  /api/recent-actions/type/{action_type}:
    get:
      description: Get a list of recent actions for a specific type
      parameters:
      - description: Action Type
        in: path
        name: action_type
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of recent actions for the type
          schema:
            items:
              $ref: '#/definitions/entity.RecentAction'
            type: array
      summary: Get recent actions by type
      tags:
      - RecentActions
  /api/recent-actions/user/{user_id}:
    get:
      description: Get a list of recent actions for a specific user
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of recent actions for the user
          schema:
            items:
              $ref: '#/definitions/entity.RecentAction'
            type: array
        "400":
          description: Invalid user ID
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get recent actions by user ID
      tags:
      - RecentActions
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"